requirements.txt
# FastAPI and Server
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic>=2.3.0,<3.0.0
pydantic-settings>=2.1.0

# Google Generative AI
google-generativeai==0.3.2

# PDF Processing
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.0.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0

# Utilities
python-dotenv==1.0.1 